+++ eras/alonzo/impl/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Version history for `cardano-ledger-alonzo`
 
+## 1.0.1.0
+
+* Add `ToJSON` instance for `AlonzoTxOut`, `AlonzoScript` and `Datum`
+
 ## 1.0.0.0
 
 * First properly versioned release.