+++ eras/shelley/impl/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Version history for `cardano-ledger-shelley`
 
+## 1.0.1.0
+
+* Add `ToJSON` instance for `ShelleyTxOut`
+
 ## 1.0.0.0
 
 * First properly versioned release.